+++ src/main/java/com/hbm/tileentity/machine/TileEntityStationPropulsionCreative.java
@@ -37,7 +37,7 @@ public class TileEntityStationPropulsionCreative extends TileEntity implements I
 	}
 
 	@Override
-	public boolean canPerformBurn(int shipMass, int deltaV) {
+	public boolean canPerformBurn(int shipMass, double deltaV) {
 		return true;
 	}
